Moved test folder out from src folder and fixed imports
