+++ __test__/dynamodb/dbHelper.test.ts
@@ -1,4 +1,4 @@
-import { removeDynamoDocumentVersion } from '../../dynamodb/dbHelper';
+import { removeDynamoDocumentVersion } from '../../src/dynamodb/dbHelper';
 import { TEST_GAME_OBJECT1 } from '../testConstants';
 
 describe('test removeGameDocumentVersion', () => {
