+++ __test__/games/mahjong/Tile/SimpleTile.test.ts
@@ -1,5 +1,5 @@
-import { SimpleTiles } from '../../../../games/mahjong/Tile/SimpleTiles';
-import { SimpleTileTypes } from '../../../../games/mahjong/Tile/types/SimpleTileTypes';
+import { SimpleTiles } from '../../../../src/games/mahjong/Tile/SimpleTiles';
+import { SimpleTileTypes } from '../../../../src/games/mahjong/Tile/types/SimpleTileTypes';
 
 test('SimpleTiles fails to initialize is range is not between 1 - 9', () => {
   const t = () => {
